extend cli to allow the exported value to also be a function that returns a Solution
